+++ frontend/src/pages/branch/index.jsx
@@ -10,21 +10,26 @@ import './branch.css';
 function Branch() {
   const navigate = useNavigate();
 
+  // Branch component
   useEffect(() => {
     GetUserData()
       .then((data) => {
         console.log(data);
 
-        if (data.user.role !== 'manager') {
-          navigate.push('/');
+        if (data.user.role === 'manager') {
+          navigate('/branch');
+          return;
+        }
 
+        if (data.user.role !== 'manager') {
+          navigate('/');
           localStorage.clear();
         }
       })
       .catch((error) => {
         console.error('Error fetching user data:', error);
       });
-  });
+  }, []); // Pass an empty dependency array
 
   return (
     <div className="page column">
